fix some warnings

+++ src/RtMidi.cpp
@@ -600,7 +600,7 @@ void MidiInCore :: openPort( unsigned int portNumber, const std::string portName
   }
 
   CFRunLoopRunInMode( kCFRunLoopDefaultMode, 0, false );
-  unsigned int nSrc = MIDIGetNumberOfSources();
+  size_t nSrc = MIDIGetNumberOfSources();
   if (nSrc < 1) {
     errorString_ = "MidiInCore::openPort: no MIDI input sources found!";
     error( RtMidiError::NO_DEVICES_FOUND, errorString_ );
@@ -690,7 +690,7 @@ void MidiInCore :: closePort( void )
 unsigned int MidiInCore :: getPortCount()
 {
   CFRunLoopRunInMode( kCFRunLoopDefaultMode, 0, false );
-  return MIDIGetNumberOfSources();
+  return (unsigned int)MIDIGetNumberOfSources();
 }
 
 // This function was submitted by Douglas Casey Tucker and apparently
@@ -774,7 +774,7 @@ static CFStringRef ConnectedEndpointName( MIDIEndpointRef endpoint )
 
   // Does the endpoint have connections?
   CFDataRef connections = NULL;
-  int nConnected = 0;
+  unsigned long nConnected = 0;
   bool anyStrings = false;
   err = MIDIObjectGetDataProperty( endpoint, kMIDIPropertyConnectionUniqueID, &connections );
   if ( connections != NULL ) {
@@ -888,7 +888,7 @@ void MidiOutCore :: initialize( const std::string& clientName )
 unsigned int MidiOutCore :: getPortCount()
 {
   CFRunLoopRunInMode( kCFRunLoopDefaultMode, 0, false );
-  return MIDIGetNumberOfDestinations();
+  return (unsigned int)MIDIGetNumberOfDestinations();
 }
 
 std::string MidiOutCore :: getPortName( unsigned int portNumber )
@@ -924,7 +924,7 @@ void MidiOutCore :: openPort( unsigned int portNumber, const std::string portNam
   }
 
   CFRunLoopRunInMode( kCFRunLoopDefaultMode, 0, false );
-  unsigned int nDest = MIDIGetNumberOfDestinations();
+  unsigned int nDest = (unsigned int)MIDIGetNumberOfDestinations();
   if (nDest < 1) {
     errorString_ = "MidiOutCore::openPort: no MIDI output destinations found!";
     error( RtMidiError::NO_DEVICES_FOUND, errorString_ );
@@ -1011,7 +1011,7 @@ void MidiOutCore :: sendMessage( std::vector<unsigned char> *message )
 {
   // We use the MIDISendSysex() function to asynchronously send sysex
   // messages.  Otherwise, we use a single CoreMidi MIDIPacket.
-  unsigned int nBytes = message->size();
+  size_t nBytes = message->size();
   if ( nBytes == 0 ) {
     errorString_ = "MidiOutCore::sendMessage: no data in message argument!";      
     error( RtMidiError::WARNING, errorString_ );
